# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "VINSight Ford API",
    "description" : "Backend SOA da plataforma **VINSight Ford**: fila de clientes priorizada pelo risco de evasão, visão 360° do cliente, passaporte do veículo e registro do desfecho de cada contato.\n\n### Como autenticar\n1. Em **Autenticação → POST /api/v1/auth/login**, clique em *Try it out* e envie o exemplo (`consultor@ford.com.br` / `consultor123`).\n2. Copie o `accessToken` da resposta.\n3. Clique em **Authorize** (cadeado), cole o token e confirme. Ele vale 15 minutos.\n\n| Usuário | Senha | Perfil | Unidade |\n|---|---|---|---|\n| consultor@ford.com.br | consultor123 | CONSULTOR | Ford Morumbi |\n| consultor.campinas@ford.com.br | consultor123 | CONSULTOR | Ford Campinas |\n| gerente@ford.com.br | gerente123 | GERENTE | Ford Morumbi |\n| analista@ford.com.br | analista123 | ANALISTA_FORD | rede inteira |\n| admin@ford.com.br | admin123 | ADMIN | rede inteira |\n\n### Erros\nTodo erro sai em **Problem Details (RFC 7807)**, com `Content-Type: application/problem+json`. Trate o erro pelo campo `type` (ex.: `https://vinsight.ford/errors/token-expirado`), nunca pelo texto. Cada endpoint lista abaixo os erros que pode responder, com um exemplo de cada.\n",
    "contact" : {
      "name" : "Grupo 02 · Glauco Heitor Gonçalves e Pedro Henrique Junqueira"
    },
    "version" : "Sprint 3"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
...
  } ],
...
  "paths" : {
    "/api/v1/vehicles/{vin}" : {
      "get" : {
        "tags" : [ "Veículos (Vehicle Service)" ],
        "summary" : "Passaporte do veículo: garantia, próxima revisão, aderência à rede, telemetria e histórico",
        "operationId" : "passaporte",
        "parameters" : [ {
          "name" : "vin",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "pattern" : "[A-HJ-NPR-Z0-9]{17}"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DadosPassaporteVeiculo"
                }
              }
            }
          },
          "422" : {
            "description" : "Erro de validação",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problema"
                },
                "examples" : {
                  "validacao" : {
                    "summary" : "Um ou mais campos estão inválidos.",
                    "value" : {
                      "type" : "https://vinsight.ford/errors/validacao",
                      "title" : "Erro de validação",
                      "status" : 422,
                      "detail" : "Um ou mais campos estão inválidos.",
                      "instance" : "/api/v1/...",
                      "timestamp" : "2026-09-22T14:30:00Z",
                      "correlationId" : "b7f3c2a1-5d4e-4f6a-9b8c-1d2e3f4a5b6c",
                      "violacoes" : [ {
                        "mensagem" : "não deve ser nulo",
                        "campo" : "desfecho"
                      } ]
                    }
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "Autenticação necessária · Token expirado",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problema"
                },
                "examples" : {
                  "nao-autenticado" : {
                    "summary" : "Envie o header Authorization: Bearer <accessToken>.",
                    "value" : {
                      "type" : "https://vinsight.ford/errors/nao-autenticado",
                      "title" : "Autenticação necessária",
                      "status" : 401,
                      "detail" : "Envie o header Authorization: Bearer <accessToken>.",
                      "instance" : "/api/v1/...",
                      "timestamp" : "2026-09-22T14:30:00Z",
                      "correlationId" : "b7f3c2a1-5d4e-4f6a-9b8c-1d2e3f4a5b6c"
                    }
                  },
                  "token-expirado" : {
                    "summary" : "Token JWT expirado.",
                    "value" : {
                      "type" : "https://vinsight.ford/errors/token-expirado",
                      "title" : "Token expirado",
                      "status" : 401,
                      "detail" : "Token JWT expirado.",
                      "instance" : "/api/v1/...",
                      "timestamp" : "2026-09-22T14:30:00Z",
                      "correlationId" : "b7f3c2a1-5d4e-4f6a-9b8c-1d2e3f4a5b6c"
                    }
                  }
                }
              }
            }
          },
          "403" : {
            "description" : "Acesso negado",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problema"
                },
                "examples" : {
                  "perfil-sem-permissao" : {
                    "summary" : "Seu perfil não tem permissão para este recurso.",
                    "value" : {
                      "type" : "https://vinsight.ford/errors/perfil-sem-permissao",
                      "title" : "Acesso negado",
                      "status" : 403,
                      "detail" : "Seu perfil não tem permissão para este recurso.",
                      "instance" : "/api/v1/...",
                      "timestamp" : "2026-09-22T14:30:00Z",
                      "correlationId" : "b7f3c2a1-5d4e-4f6a-9b8c-1d2e3f4a5b6c"
                    }
                  },
                  "outra-concessionaria" : {
                    "summary" : "Recurso pertence a outra concessionária.",
                    "value" : {
                      "type" : "https://vinsight.ford/errors/outra-concessionaria",
                      "title" : "Acesso negado",
                      "status" : 403,
                      "detail" : "Recurso pertence a outra concessionária.",
                      "instance" : "/api/v1/...",
                      "timestamp" : "2026-09-22T14:30:00Z",
                      "correlationId" : "b7f3c2a1-5d4e-4f6a-9b8c-1d2e3f4a5b6c"
                    }
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "Recurso não encontrado",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problema"
                },
                "examples" : {
                  "nao-encontrado" : {
                    "summary" : "Nenhum registro com este identificador.",
                    "value" : {
                      "type" : "https://vinsight.ford/errors/nao-encontrado",
                      "title" : "Recurso não encontrado",
                      "status" : 404,
                      "detail" : "Nenhum registro com este identificador.",
                      "instance" : "/api/v1/...",
                      "timestamp" : "2026-09-22T14:30:00Z",
                      "correlationId" : "b7f3c2a1-5d4e-4f6a-9b8c-1d2e3f4a5b6c"
                    }
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "Erro interno",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problema"
                },
                "examples" : {
                  "erro-interno" : {
                    "summary" : "Erro inesperado. Informe o correlationId ao suporte.",
                    "value" : {
                      "type" : "https://vinsight.ford/errors/erro-interno",
                      "title" : "Erro interno",
                      "status" : 500,
                      "detail" : "Erro inesperado. Informe o correlationId ao suporte.",
                      "instance" : "/api/v1/...",
                      "timestamp" : "2026-09-22T14:30:00Z",
                      "correlationId" : "b7f3c2a1-5d4e-4f6a-9b8c-1d2e3f4a5b6c"
                    }
                  }
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer-key" : [ ]
        } ]
      },
      "put" : {
        "tags" : [ "Veículos (Vehicle Service)" ],
        "summary" : "Atualiza um veículo",
        "operationId" : "atualizar",
        "parameters" : [ {
          "name" : "vin",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "pattern" : "[A-HJ-NPR-Z0-9]{17}"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DadosAtualizacaoVeiculo"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DadosDetalheVeiculo"
                }
              }
            }
          },
          "400" : {
            "description" : "Requisição inválida",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problema"
                },
                "examples" : {
                  "requisicao-invalida" : {
                    "summary" : "Corpo da requisição malformado ou ilegível.",
                    "value" : {
                      "type" : "https://vinsight.ford/errors/requisicao-invalida",
                      "title" : "Requisição inválida",
                      "status" : 400,
                      "detail" : "Corpo da requisição malformado ou ilegível.",
                      "instance" : "/api/v1/...",
                      "timestamp" : "2026-09-22T14:30:00Z",
                      "correlationId" : "b7f3c2a1-5d4e-4f6a-9b8c-1d2e3f4a5b6c"
                    }
                  }
                }
              }
            }
          },
          "422" : {
            "description" : "Erro de validação",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problema"
                },
                "examples" : {
                  "validacao" : {
                    "summary" : "Um ou mais campos estão inválidos.",
                    "value" : {
                      "type" : "https://vinsight.ford/errors/validacao",
                      "title" : "Erro de validação",
                      "status" : 422,
                      "detail" : "Um ou mais campos estão inválidos.",
                      "instance" : "/api/v1/...",
                      "timestamp" : "2026-09-22T14:30:00Z",
                      "correlationId" : "b7f3c2a1-5d4e-4f6a-9b8c-1d2e3f4a5b6c",
                      "violacoes" : [ {
                        "mensagem" : "não deve ser nulo",
                        "campo" : "desfecho"
                      } ]
                    }
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "Autenticação necessária · Token expirado",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problema"
                },
                "examples" : {
                  "nao-autenticado" : {
                    "summary" : "Envie o header Authorization: Bearer <accessToken>.",
                    "value" : {
                      "type" : "https://vinsight.ford/errors/nao-autenticado",
                      "title" : "Autenticação necessária",
                      "status" : 401,
                      "detail" : "Envie o header Authorization: Bearer <accessToken>.",
                      "instance" : "/api/v1/...",
                      "timestamp" : "2026-09-22T14:30:00Z",
                      "correlationId" : "b7f3c2a1-5d4e-4f6a-9b8c-1d2e3f4a5b6c"
                    }
                  },
                  "token-expirado" : {
                    "summary" : "Token JWT expirado.",
                    "value" : {
                      "type" : "https://vinsight.ford/errors/token-expirado",
                      "title" : "Token expirado",
                      "status" : 401,
                      "detail" : "Token JWT expirado.",
                      "instance" : "/api/v1/...",
                      "timestamp" : "2026-09-22T14:30:00Z",
                      "correlationId" : "b7f3c2a1-5d4e-4f6a-9b8c-1d2e3f4a5b6c"
                    }
                  }
                }
              }
            }
          },
          "403" : {
            "description" : "Acesso negado",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problema"
                },
                "examples" : {
                  "perfil-sem-permissao" : {
                    "summary" : "Seu perfil não tem permissão para este recurso.",
                    "value" : {
                      "type" : "https://vinsight.ford/errors/perfil-sem-permissao",
                      "title" : "Acesso negado",
                      "status" : 403,
                      "detail" : "Seu perfil não tem permissão para este recurso.",
                      "instance" : "/api/v1/...",
                      "timestamp" : "2026-09-22T14:30:00Z",
                      "correlationId" : "b7f3c2a1-5d4e-4f6a-9b8c-1d2e3f4a5b6c"
                    }
                  },
                  "outra-concessionaria" : {
                    "summary" : "Recurso pertence a outra concessionária.",
                    "value" : {
                      "type" : "https://vinsight.ford/errors/outra-concessionaria",
                      "title" : "Acesso negado",
                      "status" : 403,
                      "detail" : "Recurso pertence a outra concessionária.",
                      "instance" : "/api/v1/...",
                      "timestamp" : "2026-09-22T14:30:00Z",
                      "correlationId" : "b7f3c2a1-5d4e-4f6a-9b8c-1d2e3f4a5b6c"
                    }
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "Recurso não encontrado",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problema"
                },
                "examples" : {
                  "nao-encontrado" : {
                    "summary" : "Nenhum registro com este identificador.",
                    "value" : {
                      "type" : "https://vinsight.ford/errors/nao-encontrado",
                      "title" : "Recurso não encontrado",
                      "status" : 404,
                      "detail" : "Nenhum registro com este identificador.",
                      "instance" : "/api/v1/...",
                      "timestamp" : "2026-09-22T14:30:00Z",
                      "correlationId" : "b7f3c2a1-5d4e-4f6a-9b8c-1d2e3f4a5b6c"
                    }
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "Erro interno",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problema"
                },
                "examples" : {
                  "erro-interno" : {
                    "summary" : "Erro inesperado. Informe o correlationId ao suporte.",
                    "value" : {
                      "type" : "https://vinsight.ford/errors/erro-interno",
                      "title" : "Erro interno",
                      "status" : 500,
                      "detail" : "Erro inesperado. Informe o correlationId ao suporte.",
                      "instance" : "/api/v1/...",
                      "timestamp" : "2026-09-22T14:30:00Z",
                      "correlationId" : "b7f3c2a1-5d4e-4f6a-9b8c-1d2e3f4a5b6c"
                    }
                  }
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer-key" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "Veículos (Vehicle Service)" ],
        "summary" : "Inativa um veículo (status=INATIVO)",
        "operationId" : "inativar",
        "parameters" : [ {
          "name" : "vin",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "pattern" : "[A-HJ-NPR-Z0-9]{17}"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Removido (inativação lógica). Sem corpo."
          },
          "422" : {
            "description" : "Erro de validação",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problema"
                },
                "examples" : {
                  "validacao" : {
                    "summary" : "Um ou mais campos estão inválidos.",
                    "value" : {
                      "type" : "https://vinsight.ford/errors/validacao",
                      "title" : "Erro de validação",
                      "status" : 422,
                      "detail" : "Um ou mais campos estão inválidos.",
                      "instance" : "/api/v1/...",
                      "timestamp" : "2026-09-22T14:30:00Z",
                      "correlationId" : "b7f3c2a1-5d4e-4f6a-9b8c-1d2e3f4a5b6c",
                      "violacoes" : [ {
                        "mensagem" : "não deve ser nulo",
                        "campo" : "desfecho"
                      } ]
                    }
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "Autenticação necessária · Token expirado",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problema"
                },
                "examples" : {
                  "nao-autenticado" : {
                    "summary" : "Envie o header Authorization: Bearer <accessToken>.",
                    "value" : {
                      "type" : "https://vinsight.ford/errors/nao-autenticado",
                      "title" : "Autenticação necessária",
                      "status" : 401,
                      "detail" : "Envie o header Authorization: Bearer <accessToken>.",
                      "instance" : "/api/v1/...",
                      "timestamp" : "2026-09-22T14:30:00Z",
                      "correlationId" : "b7f3c2a1-5d4e-4f6a-9b8c-1d2e3f4a5b6c"
                    }
                  },
                  "token-expirado" : {
                    "summary" : "Token JWT expirado.",
                    "value" : {
                      "type" : "https://vinsight.ford/errors/token-expirado",
                      "title" : "Token expirado",
                      "status" : 401,
                      "detail" : "Token JWT expirado.",
                      "instance" : "/api/v1/...",
                      "timestamp" : "2026-09-22T14:30:00Z",
                      "correlationId" : "b7f3c2a1-5d4e-4f6a-9b8c-1d2e3f4a5b6c"
                    }
                  }
                }
              }
            }
          },
          "403" : {
            "description" : "Acesso negado",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problema"
                },
                "examples" : {
                  "perfil-sem-permissao" : {
                    "summary" : "Seu perfil não tem permissão para este recurso.",
                    "value" : {
                      "type" : "https://vinsight.ford/errors/perfil-sem-permissao",
                      "title" : "Acesso negado",
                      "status" : 403,
                      "detail" : "Seu perfil não tem permissão para este recurso.",
                      "instance" : "/api/v1/...",
                      "timestamp" : "2026-09-22T14:30:00Z",
                      "correlationId" : "b7f3c2a1-5d4e-4f6a-9b8c-1d2e3f4a5b6c"
                    }
                  },
                  "outra-concessionaria" : {
                    "summary" : "Recurso pertence a outra concessionária.",
                    "value" : {
                      "type" : "https://vinsight.ford/errors/outra-concessionaria",
                      "title" : "Acesso negado",
                      "status" : 403,
                      "detail" : "Recurso pertence a outra concessionária.",
                      "instance" : "/api/v1/...",
                      "timestamp" : "2026-09-22T14:30:00Z",
                      "correlationId" : "b7f3c2a1-5d4e-4f6a-9b8c-1d2e3f4a5b6c"
                    }
                  }
                }
              }
            }
          },
          "404" : {
            "description" : "Recurso não encontrado",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problema"
                },
                "examples" : {
                  "nao-encontrado" : {
                    "summary" : "Nenhum registro com este identificador.",
                    "value" : {
                      "type" : "https://vinsight.ford/errors/nao-encontrado",
                      "title" : "Recurso não encontrado",
                      "status" : 404,
                      "detail" : "Nenhum registro com este identificador.",
                      "instance" : "/api/v1/...",
                      "timestamp" : "2026-09-22T14:30:00Z",
                      "correlationId" : "b7f3c2a1-5d4e-4f6a-9b8c-1d2e3f4a5b6c"
                    }
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "Erro interno",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problema"
                },
                "examples" : {
                  "erro-interno" : {
                    "summary" : "Erro inesperado. Informe o correlationId ao suporte.",
                    "value" : {
                      "type" : "https://vinsight.ford/errors/erro-interno",
                      "title" : "Erro interno",
                      "status" : 500,
                      "detail" : "Erro inesperado. Informe o correlationId ao suporte.",
                      "instance" : "/api/v1/...",
                      "timestamp" : "2026-09-22T14:30:00Z",
                      "correlationId" : "b7f3c2a1-5d4e-4f6a-9b8c-1d2e3f4a5b6c"
                    }
                  }
                }
              }
            }
          }
        },
        "security" : [ {
          "bearer-key" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "Violacao" : {
        "type" : "object",
        "description" : "Campo inválido e o motivo",
        "properties" : {
          "campo" : {
            "type" : "string",
            "example" : "desfecho"
          },
          "mensagem" : {
            "type" : "string",
            "example" : "não deve ser nulo"
          }
        }
      },
      "Problema" : {
        "type" : "object",
        "description" : "Erro no formato Problem Details (RFC 7807)",
        "properties" : {
          "type" : {
            "type" : "string",
            "format" : "uri",
            "description" : "Identifica a situação; o app trata o erro por este campo",
            "example" : "https://vinsight.ford/errors/validacao"
          },
          "title" : {
            "type" : "string",
            "example" : "Erro de validação"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32",
            "example" : 422
          },
          "detail" : {
            "type" : "string",
            "description" : "Explicação em português, pronta para a tela",
            "example" : "Um ou mais campos estão inválidos."
          },
          "instance" : {
            "type" : "string",
            "description" : "Rota que gerou o erro",
            "example" : "/api/v1/leads/12"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time",
            "example" : "2026-09-22T14:30:00Z"
          },
          "correlationId" : {
            "type" : "string",
            "description" : "Mesmo valor do header X-Correlation-Id e da linha do log do servidor",
            "example" : "b7f3c2a1-5d4e-4f6a-9b8c-1d2e3f4a5b6c"
          },
          "violacoes" : {
            "type" : "array",
            "description" : "Só nos erros de validação (422)",
            "items" : {
              "$ref" : "#/components/schemas/Violacao"
            }
          }
        }
      },
      "DadosAtualizacaoVeiculo" : {
        "type" : "object",
        "properties" : {
          "modelo" : {
            "type" : "string"
          },
          "versao" : {
            "type" : "string"
          },
          "anoModelo" : {
            "type" : "integer",
            "format" : "int32"
          },
          "dataCompra" : {
            "type" : "string",
            "format" : "date"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "ATIVO", "INATIVO", "VENDIDO" ]
          },
          "clienteId" : {
            "type" : "integer",
            "format" : "int64"
          }
        }
      },
      "ClienteResumo" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "nome" : {
            "type" : "string"
          },
          "cpf" : {
            "type" : "string"
          }
        }
      },
      "ConcessionariaResumo" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "nomeFantasia" : {
            "type" : "string"
          }
        }
      },
      "DadosDetalheVeiculo" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "vin" : {
            "type" : "string"
          },
          "placa" : {
            "type" : "string"
          },
          "modelo" : {
            "type" : "string"
          },
          "versao" : {
            "type" : "string"
          },
          "anoFabricacao" : {
            "type" : "integer",
            "format" : "int32"
          },
          "anoModelo" : {
            "type" : "integer",
            "format" : "int32"
          },
          "dataCompra" : {
            "type" : "string",
            "format" : "date"
          },
          "status" : {
            "type" : "string",
            "enum" : [ "ATIVO", "INATIVO", "VENDIDO" ]
          },
          "cliente" : {
            "$ref" : "#/components/schemas/ClienteResumo"
          },
          "concessionariaCompra" : {
            "$ref" : "#/components/schemas/ConcessionariaResumo"
          }
        }
      },
...
      "DadosPassaporteVeiculo" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "vin" : {
            "type" : "string"
          },
          "placa" : {
            "type" : "string"
          },
          "modelo" : {
            "type" : "string"
          },
          "versao" : {
            "type" : "string"
          },
          "ano" : {
            "type" : "integer",
            "format" : "int32"
          },
          "cor" : {
            "type" : "string"
          },
          "quilometragemEstimada" : {
            "type" : "integer",
            "format" : "int32"
          },
          "garantia" : {
            "$ref" : "#/components/schemas/Garantia"
          },
          "proximaRevisaoPrevista" : {
            "$ref" : "#/components/schemas/ProximaRevisao"
          },
          "aderenciaRede" : {
            "type" : "number",
            "format" : "double"
          },
          "ultimaTelemetria" : {
            "$ref" : "#/components/schemas/Telemetria"
          },
          "historico" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ItemHistorico"
            }
          }
        }
      },
      "Garantia" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "type" : "string",
            "enum" : [ "ATIVA", "PROXIMA_DO_FIM", "ENCERRADA" ]
          },
          "dataLimite" : {
            "type" : "string",
            "format" : "date"
          },
          "mesesRestantes" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "ItemHistorico" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "data" : {
            "type" : "string",
            "format" : "date"
          },
          "tipoServico" : {
            "type" : "string",
            "enum" : [ "REVISAO_PROGRAMADA", "TROCA_OLEO", "REPARO", "GARANTIA", "RECALL" ]
          },
          "descricao" : {
            "type" : "string"
          },
          "valor" : {
            "type" : "number"
          },
          "concessionaria" : {
            "type" : "string"
          },
          "naRede" : {
            "type" : "boolean"
          }
        }
      },
      "ProximaRevisao" : {
        "type" : "object",
        "properties" : {
          "km" : {
            "type" : "integer",
            "format" : "int32"
          },
          "dataEstimada" : {
            "type" : "string",
            "format" : "date"
          },
          "situacao" : {
            "type" : "string",
            "enum" : [ "EM_DIA", "PROXIMA", "VENCIDA" ]
          }
        }
      },
      "Telemetria" : {
        "type" : "object",
        "properties" : {
          "recebidaEm" : {
            "type" : "string",
            "format" : "date-time"
          },
          "codigosFalha" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearer-key" : {
        "type" : "http",
        "description" : "Access token devolvido por POST /api/v1/auth/login (válido por 15 minutos).",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}